+++ src/routes.ts
@@ -1,10 +1,14 @@
 import {Router, Request, Response} from 'express';
 
+import { CreateUserController } from './controllers/user/CreateUserController';
+import { CreateUserAdminController } from './controllers/userAdmin/CreateUserAdminController';
 
 const router = Router();
+// Routes users/client
+router.post('/user/client', new CreateUserController().handle)
 
-router.get('/teste', (req: Request, res: Response) => {
-    return res.json({nome: "clovis"})
-})
+// Routes users/admin
+
+router.post('/user/admin', new CreateUserAdminController().handle)
 
 export { router };