+++ src/services/userAdmin/CreateUserAdminService.ts
@@ -1,0 +1,13 @@
+interface userAdminRequest{
+    name: string;
+    email: string;
+    password: string;
+    contact: string;
+    local: string;
+}
+class CreateUserAdminService{
+    async execute({name, email, password, contact, local}: userAdminRequest){
+        return {name: name}
+    }
+}
+export { CreateUserAdminService }