+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added cors
 - Added Prisma
 - Created tables
+- Adjusted tables
+- Created controller and user service
 
 ### Fixed